working circom and spartan2 pipeline
