+++ Cargo.toml
@@ -40,7 +40,6 @@ ark-serialize = { version = "0.4.2", default-features = false, features = [
     "derive",
 ] }
 
-# ark-bls12-381 = { version = "^0.4.0", default-features = false, features = [ "curve" ] }
 criterion = { version = "0.3.1", features = ["html_reports"] }
 num-integer = "0.1.45"
 seq-macro = "0.3.3"
@@ -53,16 +52,8 @@ hashbrown = "0.14.0"
 ff = "0.13.0"
 pasta_curves = "0.5.1"
 bellpepper-core = "0.2.1"
-spartan2 = { git = "https://github.com/microsoft/Spartan2.git" }
-# spartan2 = { path = "./spartan2/" }
-bellperson = { version = "0.25", default-features = false }
-# circom-scotia = { path = "./circom-scotia/" }
 circom-scotia = "0.1.2"
-nova-scotia = "0.5.0"
-num-bigint = "0.4.4"
-num-traits = "0.2.16"
-serde = "1.0.188"
-serde_json = "1.0.107"
+spartan2 = "0.1.0"
 
 [dev-dependencies]
 criterion = "0.3.1"
